requirements.txt
django
djangorestframework
django-cors-headers
Pillow
django-database-url
sqlparse
gunicorn
psycopg2